pipeline/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.15
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
azure-ai-documentintelligence==1.0.2
azure-core==1.32.0
azure-cosmos==4.9.0
azure-functions==1.21.3
azure-functions-durable==1.2.10
azure-identity==1.21.0
azure-storage-blob==12.25.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
furl==2.1.4
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.9.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.3.0
openai==1.70.0
orderedmultidict==1.0.1
propcache==0.3.1
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==2.3.0
yarl==1.18.3
